Adds dummy test for module chip utils
