+++ test/components/modulplanner/modulechip/util.test.js
@@ -1,0 +1,5 @@
+import {depsMatter} from "../../../../src/components/modulplanner/modulechip/util";
+
+test("depsMatter dummy", () => {
+  expect(depsMatter({s: []}, "test")).toBe(false);
+});
